simple_requirements.txt
# Simple WebRTC AI Assistant Requirements

fastapi>=0.115.6
uvicorn>=0.32.0
python-dotenv>=1.0.0
loguru>=0.7.3
pydantic>=2.10.6

# Pipecat AI framework with required dependencies
pipecat-ai[runner,webrtc,openai,deepgram,cartesia,silero]
